requirements.txt
amqp==2.6.0
arrow==0.15.8
asgiref==3.2.10
billiard==3.6.3.0
celery==4.4.6
click==7.1.2
Django==3.0.8
django-cors-headers==3.4.0
django-filter==2.3.0
django-templated-mail==1.1.1
djangorestframework==3.11.0
djangorestframework-jsonapi==3.1.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.4.0
djoser==2.0.3
future==0.18.2
inflection==0.5.0
kombu==4.6.11
psycopg2-binary==2.8.5
PyJWT==1.7.1
python-dateutil==2.8.1
python-dotenv==0.14.0
pytz==2020.1
redis==3.5.3
six==1.15.0
sqlparse==0.3.1
vine==1.3.0
